feat: init repo

+++ src/main/java/com/lcm/food/application/domain/vos/CustomerVO.java
@@ -1,4 +1,4 @@
 package com.lcm.food.application.domain.vos;
 
-public record CustomerVO(Integer ID) {
+public record CustomerVO(Long ID) {
 }
